Refactor get model command (excluding version history)
